--- src/main/java/baseball/Controller/GameController.java
@@ -1,19 +1,0 @@
-package baseball.Controller;
-
-import baseball.Model.AnswerNumber;
-import baseball.Model.UserNumber;
-import baseball.Service.NumberMatcher;
-import baseball.View.InputView;
-import baseball.View.OutputView;
-
-public class GameController {
-
-    private AnswerNumber answerNumber;
-    private UserNumber userNumber;
-    private NumberMatcher numberMatcher;
-    private InputView inputView;
-    private OutputView outputView;
-
-
-
-}
