+++ src/main/java/baseball/model/AnswerNumber.java
@@ -1,4 +1,4 @@
-package baseball.Model;
+package baseball.model;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
