+++ src/main/java/baseball/model/UserNumber.java
@@ -1,4 +1,4 @@
-package baseball.Model;
+package baseball.model;
 
 import java.util.ArrayList;
 import java.util.HashSet;
