+++ src/main/java/baseball/view/InputView.java
@@ -1,4 +1,4 @@
-package baseball.View;
+package baseball.view;
 
 import camp.nextstep.edu.missionutils.Console;
